Struct lazybox::ecs::entity::Entities

pub struct Entities { /* fields omitted */ }

Methods

impl Entities

fn new() -> Self

fn create(&self) -> Entity

Creates an entity

The entity is not considered alive until spawn is called

fn spawn_later(&self, entity: Entity) -> EntityRef

Spawns an entity when the state will be commited.

fn spawn(&mut self, entity: Entity) -> EntityRef

Spawns an entity making it alive

fn remove_later<'a>(&self, accessor: Accessor<'a>)

Schedule the removal of an entity.

The entity will be removed at the next state commit

fn iter(&self) -> Iter

Creates a new entity iterator

fn entity_ref<'a>(&self, accessor: Accessor<'a>) -> EntityRef

Creates an entity reference

fn upgrade(&self, entity_ref: EntityRef) -> Option<Accessor>

Uprades an EntityRef to an Accessor

Returns None if the Entity has been killed

fn push_removes(&mut self) -> Vec<Entity>

Removes all killed entities and returns them.

The entities returned might be recycled in the future and need to be used with care.

fn commit(&mut self)

Commit the entities changes.