Module lazybox::ecs::policy

Defines some compile-time configuration

A policy is a way to define how do you want to represents your entity Id et Version in memory. You can choose between different unsigned interger sizes that will determine a maximum of entities that can exists at the same time.

Functions

id_from_usize
Converts a usize to an Id.

max_entity_count
Returns the maxmimum of entities that can exists at the same time

Type Definitions

Id
The id type for an entity

IdSet

Version
The version type for an entity