Module lazybox::maths

Reexports

pub use cgmath::*;

Macros

assert_relative_eq

assert_relative_ne

assert_ulps_eq

assert_ulps_ne

relative_eq
Predicate for testing the approximate equality of two values.

relative_ne
Predicate for testing the approximate inequality of two values.

ulps_eq
Predicate for testing the approximate equality of two values using a maximum ULPs (Units in Last Place).

ulps_ne
Predicate for testing the approximate inequality of two values using a maximum ULPs (Units in Last Place).