Struct lazybox::modules::storages::packed::Packed

pub struct Packed<V> { /* fields omitted */ }

A Storage that hold its values in a contiguous vector.

Methods

impl<V> Packed<V>

fn new() -> Packed<V>

Constructs a new empty Packed` storage.

fn insert<'a>(&mut self, key: Accessor<'a>, component: V) -> bool

Associate a new Component V to the entity

fn remove<'a>(&mut self, key: Accessor<'a>)

Detach a Component V from the entity

fn with_capacity(capacity: usize) -> Packed<V>

Constructs a new empty Packed<V> with the given capacity

fn contains<'a>(&self, key: Accessor<'a>) -> bool

Returns true if a component is associated to the entity

fn get<'a>(&self, key: Accessor<'a>) -> Option<&V>

Returns a immutable access to the associated component

fn get_mut<'a>(&mut self, key: Accessor<'a>) -> Option<&mut V>

Returns a mutable access to the associated component

fn iter(&self) -> Iter<V>

An iterator visiting all component-entity pairs in arbitrary order.

fn iter_mut(&mut self) -> IterMut<V>

An iterator visiting all component-entity pairs in arbitrary order.